+++ packages/babel-preset-app-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inventorjs/babel-preset-app-react
 
+## 5.0.17
+
+### Patch Changes
+
+- change arguments to params
+
 ## 5.0.16
 
 ### Patch Changes