+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @inventorjs/cli
 
+## 5.0.17
+
+### Patch Changes
+
+- change arguments to params
+- Updated dependencies
+  - @inventorjs/cli-core@5.0.17
+  - @inventorjs/inventor-plugin-app@5.0.17
+  - @inventorjs/inventor-plugin-plugin@5.0.17
+
 ## 5.0.16
 
 ### Patch Changes